Fix: Declare namespaced functions in separate files
