+++ src/Eris/Antecedent/PrintableCharacter.php
@@ -3,16 +3,6 @@ namespace Eris\Antecedent;
 
 use Eris\Antecedent;
 
-function printableCharacter()
-{
-    return new PrintableCharacter();
-}
-
-function printableCharacters()
-{
-    return new PrintableCharacter();
-}
-
 class PrintableCharacter implements Antecedent
 {
     /**
