+++ src/Eris/Antecedent/functions.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Eris\Antecedent;
+
+function printableCharacter()
+{
+    return new PrintableCharacter();
+}
+
+function printableCharacters()
+{
+    return new PrintableCharacter();
+}
