+++ src/Eris/Generator/BooleanGenerator.php
@@ -4,11 +4,6 @@ namespace Eris\Generator;
 use Eris\Generator;
 use DomainException;
 
-function bool()
-{
-    return new BooleanGenerator();
-}
-
 class BooleanGenerator implements Generator
 {
     public function __invoke($_size)
