+++ src/Eris/Generator/ChooseGenerator.php
@@ -9,20 +9,6 @@ if (!defined('ERIS_PHP_INT_MIN')) {
     define('ERIS_PHP_INT_MIN', ~PHP_INT_MAX);
 }
 
-/**
- * Generates a number in the range from the lower bound to the upper bound,
- * inclusive. The result shrinks towards smaller absolute values.
- * The order of the parameters does not care since they are re-ordered by the
- * generator itself.
- *
- * @param $x int One of the 2 boundaries of the range
- * @param $y int The other boundary of the range
- * @return Generator\ChooseGenerator
- */
-function choose($lowerLimit, $upperLimit) {
-    return new ChooseGenerator($lowerLimit, $upperLimit);
-}
-
 class ChooseGenerator implements Generator
 {
     private $lowerLimit;
