+++ src/Eris/Generator/OneOfGenerator.php
@@ -3,14 +3,6 @@ namespace Eris\Generator;
 
 use Eris\Generator;
 
-/**
- * @return OneOfGenerator
- */
-function oneOf(/*$a, $b, ...*/)
-{
-    return new OneOfGenerator(func_get_args());
-}
-
 class OneOfGenerator implements Generator
 {
     private $generator;
