+++ src/Eris/Generator/RegexGenerator.php
@@ -8,19 +8,6 @@ use ReverseRegex\Random\SimpleRandom;
 use ReverseRegex\Parser;
 use ReverseRegex\Generator\Scope;
 
-/**
- * Note * and + modifiers cause an unbounded number of character to be generated
- * (up to plus infinity) and as such they are not supported.
- * Please use {1,N} and {0,N} instead of + and *.
- *
- * @param string $expression
- * @return Generator\RegexGenerator
- */
-function regex($expression)
-{
-    return new RegexGenerator($expression);
-}
-
 class RegexGenerator implements Generator
 {
     private $expression;
